Move ClipboardButton to dedicated file
Make room in index for React implementation
