+++ app/javascript/packages/clipboard-button/clipboard-button-element.ts
